Update NeuroML.yaml
A slight update. But, this is a bit of a strange one: I think NeuroML is a format first, and (somewhat confusingly) a tool second (and isn't that simulator actually called LEMS?). Some tools feature NML as language already, which is currently not used for the relations field. If we add file formats here, it might be good to use that field too to create links.